feat: integrate trip detail view into dashboard
- Convert TripDetail page to TripDetailView component
- Add trip viewing within dashboard layout
- Maintain persistent sidebar navigation
- Remove separate trip detail route
- Update TripCard to use dashboard navigation
